:tada: ts rewrite
