+++ package.json
@@ -23,5 +23,8 @@
   "license": "MIT",
   "dependencies": {
     "node-fetch": "^2.6.1"
+  },
+  "devDependencies": {
+    "@types/node-fetch": "^2.5.10"
   }
 }
